chore: minor ui fixes
