+++ app/package.json
@@ -16,7 +16,6 @@
     "@codemirror/language": "^6.10.1",
     "@codemirror/legacy-modes": "6.4.0",
     "@codemirror/view": "6.29.0",
-    "zustand":"5.0.3",
     "@radix-ui/react-dialog": "^1.1.6",
     "@radix-ui/react-dropdown-menu": "^2.1.6",
     "@radix-ui/react-label": "^2.1.2",
