+++ app/src/components/Search.tsx
@@ -1,5 +1,5 @@
 import { Input } from "./ui/input";
-import { useStore } from "../store";
+import { useStore } from "@/store";
 import { Grid, List, SearchIcon, XIcon } from "lucide-react";
 import { Button } from "./ui/button";
 import { Popover, PopoverContent, PopoverTrigger } from "./ui/popover";
@@ -16,8 +16,10 @@ import { Check, ChevronsUpDown } from "lucide-react";
 import React from "react";
 import { Tabs, TabsList, TabsTrigger } from "./ui/tabs";
 import SelectedTags from "./SelectedTags";
+
 const Search = () => {
-  const { templates, searchQuery, setSearchQuery, setView } = useStore();
+  const { templates, searchQuery, setSearchQuery, setView, templatesCount } =
+    useStore();
   const selectedTags = useStore((state) => state.selectedTags);
   const addSelectedTag = useStore((state) => state.addSelectedTag);
   const removeSelectedTag = useStore((state) => state.removeSelectedTag);
@@ -43,11 +45,20 @@ const Search = () => {
 
   return (
     <div className=" mx-auto p-12  border-b w-full">
-      <h1 className="text-2xl md:text-3xl xl:text-4xl font-bold text-center mb-8">
+      {/* <h1 className="text-2xl md:text-3xl xl:text-4xl font-bold text-center mb-8">
         Available Templates ({templates?.length || 0})
-      </h1>
+      </h1> */}
       <div className="max-w-xl mx-auto flex flex-col gap-2">
         <div className="relative w-full">
+          <div className="mb-2 flex flex-row gap-1">
+            <div className="text-sm text-muted-foreground">
+              Available Templates
+            </div>
+            <div className="text-sm font-bold">
+              {(templatesCount && templatesCount) || 0}
+            </div>
+          </div>
+
           <Input
             type="text"
             placeholder="Search templates..."
@@ -57,10 +68,10 @@ const Search = () => {
           />
           {searchQuery.length > 0 ? (
             <div className="cursor-pointer" onClick={() => setSearchQuery("")}>
-              <XIcon className="absolute right-3 top-1/2 -translate-y-1/2 h-5 w-5 text-gray-400" />
+              <XIcon className="absolute end-3 translate-y-3.5 top-1/2 h-5 w-5 text-gray-400" />
             </div>
           ) : (
-            <SearchIcon className="absolute right-3 top-1/2 -translate-y-1/2 h-5 w-5 text-gray-400" />
+            <SearchIcon className="absolute end-3 translate-y-3.5 top-1/2 h-5 w-5 text-gray-400" />
           )}
         </div>
 
@@ -72,10 +83,10 @@ const Search = () => {
             }}
           >
             <TabsList>
-              <TabsTrigger value="grid">
+              <TabsTrigger value="grid" className="cursor-pointer">
                 <Grid /> <span className="text-xs p-1">Grid</span>
               </TabsTrigger>
-              <TabsTrigger value="rows">
+              <TabsTrigger value="rows" className="cursor-pointer">
                 <List /> <span className="text-xs p-1">List</span>
               </TabsTrigger>
             </TabsList>
