+++ app/src/components/SelectedTags.tsx
@@ -1,5 +1,5 @@
 import { XIcon } from "lucide-react";
-import { useStore } from "../store";
+import { useStore } from "@/store";
 import { Badge } from "./ui/badge";
 
 const SelectedTags = () => {
