+++ app/src/components/TemplateGrid.tsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useState } from "react";
-import { Input } from "./ui/input";
 import {
   Card,
   CardHeader,
@@ -7,23 +6,10 @@ import {
   CardContent,
   CardFooter,
 } from "./ui/card";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-} from "./ui/dialog";
-import { Button } from "./ui/button";
-import { toast } from "sonner";
-import copy from "copy-to-clipboard";
-import { CodeEditor } from "./ui/code-editor";
-import { useStore } from "../store";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "./ui/tabs";
-import { Label } from "./ui/label";
-import { Clipboard } from "lucide-react";
+import { useStore } from "@/store";
 import { Skeleton } from "./ui/skeleton";
 import { cn } from "@/lib/utils";
+import TemplateDialog from "./TemplateDialog";
 
 interface Template {
   id: string;
@@ -49,7 +35,13 @@ interface TemplateGridProps {
 }
 
 const TemplateGrid: React.FC<TemplateGridProps> = ({ view }) => {
-  const { templates, setTemplates } = useStore();
+  const {
+    templates,
+    setTemplates,
+    setTemplatesCount,
+    filteredTemplates,
+    setFilteredTemplates,
+  } = useStore();
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
   const searchQuery = useStore((state) => state.searchQuery);
@@ -73,6 +65,8 @@ const TemplateGrid: React.FC<TemplateGridProps> = ({ view }) => {
         }
         const data = await response.json();
         setTemplates(data);
+        setFilteredTemplates(data);
+        setTemplatesCount(data.length);
         setLoading(false);
       } catch (err) {
         setError(err instanceof Error ? err.message : "An error occurred");
@@ -81,7 +75,7 @@ const TemplateGrid: React.FC<TemplateGridProps> = ({ view }) => {
     };
 
     fetchTemplates();
-  }, [setTemplates]);
+  }, [setTemplates, setFilteredTemplates]);
 
   const fetchTemplateFiles = async (templateId: string) => {
     setModalLoading(true);
@@ -111,32 +105,25 @@ const TemplateGrid: React.FC<TemplateGridProps> = ({ view }) => {
     fetchTemplateFiles(template.id);
   };
 
-  const filteredTemplates = templates.filter((template) => {
-    // Filter by search query
-    const matchesSearch = template.name
-      .toLowerCase()
-      .includes(searchQuery.toLowerCase());
+  useEffect(() => {
+    const filtered = templates.filter((template) => {
+      // Filter by search query
+      const matchesSearch = template.name
+        .toLowerCase()
+        .includes(searchQuery.toLowerCase());
 
-    // Filter by selected tags
-    const matchesTags =
-      selectedTags.length === 0 ||
-      selectedTags.every((tag) => template.tags.includes(tag));
+      // Filter by selected tags
+      const matchesTags =
+        selectedTags.length === 0 ||
+        selectedTags.every((tag) => template.tags.includes(tag));
 
-    return matchesSearch && matchesTags;
-  });
+      return matchesSearch && matchesTags;
+    });
 
-  console.log(filteredTemplates);
-
-  const getBase64Config = () => {
-    if (!templateFiles?.dockerCompose && !templateFiles?.config) return "";
-
-    const configObj = {
-      compose: templateFiles.dockerCompose || "",
-      config: templateFiles.config || "",
-    };
-
-    return btoa(JSON.stringify(configObj, null, 2));
-  };
+    console.log("ffiltered tem", filtered.length);
+    setTemplatesCount(filtered.length);
+    setFilteredTemplates(filtered);
+  }, [searchQuery, selectedTags]);
 
   if (loading) {
     return (
@@ -192,21 +179,28 @@ const TemplateGrid: React.FC<TemplateGridProps> = ({ view }) => {
                   " cursor-pointer hover:shadow-lg transition-all duration-200 h-full max-h-[300px]",
                   {
                     "flex-col": view === "grid",
-                    "flex-row": view === "rows",
+                    "flex-row gap-0": view === "rows",
                   }
                 )}
               >
-                <CardHeader>
-                  <CardTitle className="text-xl ">
-                    <img
-                      src={`/blueprints/${template.id}/${template.logo}`}
-                      alt={template.name}
-                      className="w-12 h-12 object-contain mb-2"
-                    />
-                    {template.name}
-                  </CardTitle>
+                <CardHeader
+                  className={cn("flex  gap-2 ", {
+                    "flex-row": view === "grid",
+                    "flex-col justify-center items-center ms-4":
+                      view === "rows",
+                  })}
+                >
+                  <img
+                    src={`/blueprints/${template.id}/${template.logo}`}
+                    alt={template.name}
+                    className={cn("w-auto h-12 s object-contain mb-2", {
+                      "w-auto h-12": view === "grid",
+                      "w-12 h-auto": view === "rows",
+                    })}
+                  />
                 </CardHeader>
                 <CardContent className="flex-1">
+                  <CardTitle className="text-xl ">{template.name}</CardTitle>
                   <p className="text-sm text-muted-foreground line-clamp-2">
                     {template.description}
                   </p>
@@ -255,197 +249,12 @@ const TemplateGrid: React.FC<TemplateGridProps> = ({ view }) => {
         </div>
       </div>
 
-      <Dialog
-        open={!!selectedTemplate}
-        onOpenChange={() => setSelectedTemplate(null)}
-      >
-        <DialogContent className="!max-w-[90vw] w-full lg:max-w-[90vw] max-h-[85vh] overflow-y-auto p-0">
-          <DialogHeader className="space-y-4 border-b sticky top-0 p-4 bg-background z-10">
-            <div className="flex items-center gap-4">
-              {selectedTemplate?.logo && (
-                <img
-                  src={`/blueprints/${selectedTemplate.id}/${selectedTemplate.logo}`}
-                  alt={selectedTemplate.name}
-                  className="w-12 h-12 object-contain"
-                />
-              )}
-              <div>
-                <DialogTitle className="text-2xl">
-                  {selectedTemplate?.name}
-                </DialogTitle>
-                <div className="flex items-center gap-2 mt-1">
-                  <span className="text-sm text-gray-500">
-                    {selectedTemplate?.version}
-                  </span>
-                  <div className="flex gap-2">
-                    {selectedTemplate?.links.github && (
-                      <a
-                        href={selectedTemplate.links.github}
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        className="text-gray-600 hover:text-gray-900"
-                      >
-                        GitHub
-                      </a>
-                    )}
-                    {selectedTemplate?.links.docs && (
-                      <a
-                        href={selectedTemplate.links.docs}
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        className="text-gray-600 hover:text-gray-900"
-                      >
-                        Docs
-                      </a>
-                    )}
-
-                    <a
-                      href={`https://github.com/Dokploy/templates/tree/main/blueprints/${selectedTemplate?.id}`}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      className="text-gray-600 hover:text-gray-900"
-                    >
-                      Edit Template
-                    </a>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </DialogHeader>
-
-          <div className="p-6 pt-3 max-w-[100%] flex flex-col gap-2">
-            <DialogDescription className="text-base">
-              {selectedTemplate?.description}
-            </DialogDescription>
-            <div className="flex flex-wrap gap-1">
-              {selectedTemplate?.tags.map((tag) => (
-                <span
-                  key={tag}
-                  className="inline-flex items-center px-2 py-1 rounded-md text-xs font-medium bg-blue-100 text-blue-800"
-                >
-                  {tag}
-                </span>
-              ))}
-            </div>
-
-            {modalLoading ? (
-              <div className="py-12 text-center">
-                <div className="inline-block animate-spin rounded-full h-10 w-10 border-4 border-solid border-blue-500 border-r-transparent"></div>
-                <p className="mt-4 text-gray-600">Loading template files...</p>
-              </div>
-            ) : (
-              <div className="grid gap-4 mt-6">
-                {(templateFiles?.dockerCompose || templateFiles?.config) && (
-                  <div className="flex flex-col gap-3">
-                    <Label className=" flex  flex-col items-start w-fit justify-start gap-1">
-                      <span className="leading-tight text-xl font-semibold">
-                        Base64 Configuration
-                      </span>
-                      <span className="leading-tight text-sm  text-gray-500">
-                        Encoded template file
-                      </span>
-                    </Label>
-                    <div className="relative">
-                      <Button
-                        // variant={"outline"}
-                        className="absolute end-0
-                     "
-                        size={"icon"}
-                        onClick={() => {
-                          toast.success("Copied to clipboard");
-                          copy(getBase64Config());
-                        }}
-                      >
-                        <Clipboard />
-                      </Button>
-                      <Input
-                        value={getBase64Config()}
-                        className="max-w-6xl w-full pe-10"
-                      />
-                    </div>
-                  </div>
-                )}
-                <Tabs defaultValue="docker-compose">
-                  <TabsList>
-                    <TabsTrigger value="docker-compose">
-                      Docker Compose
-                    </TabsTrigger>
-                    <TabsTrigger value="config">Configuration</TabsTrigger>
-                  </TabsList>
-                  <TabsContent value="docker-compose">
-                    {templateFiles?.dockerCompose && (
-                      <div className=" flex flex-col relative">
-                        <Label className=" flex mb-2 flex-col items-start w-fit justify-start gap-1">
-                          <span className="leading-tight text-xl font-semibold">
-                            Docker Compose
-                          </span>
-                          <span className="leading-tight text-sm  text-gray-500">
-                            docker-compose.yml
-                          </span>
-                        </Label>
-
-                        <CodeEditor
-                          value={templateFiles.dockerCompose || ""}
-                          language="yaml"
-                          className="font-mono"
-                        />
-                        <Button
-                          onClick={() => {
-                            toast.success("Copied to clipboard");
-                            copy(templateFiles.dockerCompose || "");
-                          }}
-                          className="absolute top-10 right-2 px-3 py-1  text-sm cursor-pointer"
-                        >
-                          Copy
-                        </Button>
-                      </div>
-                    )}
-                  </TabsContent>
-
-                  <TabsContent value="config">
-                    {templateFiles?.config && (
-                      <div className="max-w-6xl w-full relative">
-                        <Label className=" flex mb-2 flex-col items-start w-fit justify-start gap-1">
-                          <span className="leading-tight text-xl font-semibold">
-                            Configuration
-                          </span>
-                          <span className="leading-tight text-sm  text-gray-500">
-                            template.yml
-                          </span>
-                        </Label>
-
-                        <CodeEditor
-                          value={templateFiles.config || ""}
-                          language="yaml"
-                          className="font-mono"
-                        />
-
-                        <Button
-                          onClick={() => {
-                            toast.success("Copied to clipboard");
-                            copy(templateFiles.config || "");
-                          }}
-                          className="absolute top-10 right-2 px-3 py-1  text-sm cursor-pointer"
-                        >
-                          Copy
-                        </Button>
-                      </div>
-                    )}
-                  </TabsContent>
-                </Tabs>
-
-                {!templateFiles?.dockerCompose && !templateFiles?.config && (
-                  <div className="text-center py-8">
-                    <p className="text-gray-500">
-                      No configuration files available for this template.
-                    </p>
-                  </div>
-                )}
-              </div>
-            )}
-          </div>
-        </DialogContent>
-      </Dialog>
+      <TemplateDialog
+        selectedTemplate={selectedTemplate}
+        templateFiles={templateFiles}
+        modalLoading={modalLoading}
+        onOpenChange={(open) => !open && setSelectedTemplate(null)}
+      />
     </>
   );
 };
